:recycle: [Refactor] useComment 분리 #main
